Mostly Blender Work.........
I worked a lot on the tiling in Unity, but no luck so far. I started creating the textures in Blender to then import it to unity. Also spent countless hours on it and then watched hundreds of tutorials to import the material that I created from blender to unity, but it did not work no matter what I did. I did some renders to show how it should look when complete. I also cut out new pieces of psyche and made them look more real by applying several modifiers to it.